+++ pmd-core/src/main/java/net/sourceforge/pmd/PmdAnalysis.java
@@ -4,6 +4,7 @@
 
 package net.sourceforge.pmd;
 
+import static net.sourceforge.pmd.lang.document.InternalApiBridge.newCollector;
 import static net.sourceforge.pmd.lang.rule.InternalApiBridge.loadRuleSetsWithoutException;
 import static net.sourceforge.pmd.lang.rule.InternalApiBridge.ruleSetApplies;
 import static net.sourceforge.pmd.util.CollectionUtil.listOf;
@@ -159,7 +160,7 @@ public final class PmdAnalysis implements AutoCloseable {
     private PmdAnalysis(PMDConfiguration config) {
         this.configuration = config;
         this.reporter = config.getReporter();
-        this.collector = FileCollector.newCollector(
+        this.collector = newCollector(
             config.getLanguageVersionDiscoverer(),
             reporter
         );